added testing for datetime and invalid inputs
