+++ app/models/task.py
@@ -1,12 +1,13 @@
 from app import db
 from sqlalchemy.orm import relationship
-
+from flask import make_response, abort
+from datetime import datetime
 
 class Task(db.Model):
     task_id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String, nullable=False)
     description = db.Column(db.String, nullable=False)
-    completed_at = db.Column(db.String, nullable=True)
+    completed_at = db.Column(db.DateTime, nullable=True)
     goal_id = db.Column(db.Integer, db.ForeignKey("goal.goal_id"), nullable = True)
     goal = relationship("Goal", back_populates="tasks")
 
@@ -29,11 +30,19 @@ class Task(db.Model):
                    }
     @classmethod
     def from_dict(cls, data):
+        if "datetime" not in data or "datetime" == None:
+            data["datetime"] = None
+        else:
+            try:
+                data["datetime"] = datetime.strptime(data["datetime"], '%m/%d/%y')
+            except:
+                abort(make_response({"details":f"datetime invalid, needs to be in form 'm/d/yy'"}, 400))
+
         if "description" in data and "title" in data:
             return Task(
                 title = data["title"],
                 description = data["description"],
-                completed_at = None
+                completed_at = data["datetime"]
             )
         else:
             return False
