+++ practice.py
@@ -1,14 +1,8 @@
-import requests
-import os
-from dotenv import load_dotenv
+from datetime import date, datetime
+# today = date.today()
+# today_str = today.strftime("%m/%d/%Y")
+# print(date.today())
 
-load_dotenv()
-def slack_send_message(task):
-    SLACK_URL = "https://slack.com/api/chat.postMessage"
-    PARAMS = {"channel":"task-notifications",
-            "text":f"Someone just completed the task {task.to_dict()['title']}",}
-    HEADERS ={
-            "Authorization": f"Bearer {os.environ.get('SLACK_BOT_TOKEN')}"
-            }
-    r = requests.post(url = SLACK_URL, headers= HEADERS, params=PARAMS)
-    print(r)
+today = "11/9/1a"
+ftoday = datetime.strptime(today, '%m/%d/%y')
+print(ftoday)
